Rework travis build to use test-omero
Major changes:

 * All .travis.yml commands now live in test-omero
 * All actions take place in docker rather than locally
 * All classes live in omero_cli_render to prevent shadowing

The next step here will be to remove the copied cli.py module
in favor of pulling the clisplit image.
