--- travis-build
@@ -1,25 +1,0 @@
-#! /bin/bash
-
-set -e
-set -u
-set -x
-
-
-OMERO_VERSION=${OMERO_VERSION:-5.4.0}
-
-python setup.py sdist
-pip install dist/*.tar.gz
-pip install omego
-
-# Install the server
-omego install --managedb --dbhost localhost --dbname omero --prestartfile $HOME/config.omero -v --release $OMERO_VERSION --ice 3.5 --no-web
-
-# Check that the plugin is installed
-OMERO.server/bin/omero render -h;
-
-
-# run the test
-cp src/omero/plugins/render.py OMERO.server/lib/python/omero/plugins/render.py
-VALUE=$(readlink -f OMERO.server)
-export PYTHONPATH=$VALUE/lib/python
-python setup.py test -t test/integration/clitest -i $VALUE/etc/ice.config -v
